Refactored whole static volt system to deal in milli volts, thus allowing for more appropriate ratios with RF. Exported a TON of configuration values (almost all machine values) to the server config. Shifted some more values there too. Getting ready for a real release after final testing!

+++ src/main/java/theking530/api/power/ItemStackStaticVoltCapability.java
@@ -15,7 +15,7 @@ public class ItemStackStaticVoltCapability extends StaticVoltHandler implements 
 	protected final PowerEnergyInterface energyInterface;
 	protected ItemStackMultiCapabilityProvider owningProvider;
 
-	public ItemStackStaticVoltCapability(String name, ItemStack container, int capacity, int maxInput, int maxOutput) {
+	public ItemStackStaticVoltCapability(String name, ItemStack container, long capacity, long maxInput, long maxOutput) {
 		super(capacity, maxInput, maxOutput);
 		this.name = name;
 		this.container = container;
